+++ app/config/index.js
@@ -1,0 +1,5 @@
+if (__ENVIRONMENT__ === 'production') {
+    module.exports = require('./production')
+} else {
+    module.exports = require('./default')
+}