+++ app/index.js
@@ -1,7 +1,2 @@
 import 'isomorphic-fetch';
-
-if (CONFIG.mockApi) {
-	require('./mocks');
-}
-
 export serverMiddleware from './Router';