+++ app/reducers/user.js
@@ -22,10 +22,9 @@ export default function user(state = {}, action) {
 			});
 		case USER_FETCHED:
 			return Object.assign({}, state, {
-				[action.userId]: {
-					readyState: USER_FETCHED,
-					data: action.result
-				}
+				[action.userId]: Object.assign({
+					readyState: USER_FETCHED
+				}, action.result)
 			});
 		default:
 			return state;