+++ webpack.config.js
@@ -6,12 +6,12 @@ var webpack = require('webpack');
 module.exports = {
 	devtool: '#source-map',
 	entry: [
-		'webpack-hot-middleware/client?path=/__webpack_hmr',
+		'webpack-hot-middleware/client',
 		'./app/index.js'
 	],
 	output: {
 		path: path.join(__dirname, 'dist'),
-		filename: 'bundle.js'
+		filename: 'app.js'
 	},
 	plugins: [
 		new webpack.optimize.OccurenceOrderPlugin(),
@@ -28,14 +28,6 @@ module.exports = {
 					['transform-object-assign']
 				]
 			}
-		}, {
-			test: /\.scss$/,
-			loaders: ['style', 'css?sourceMap', 'sass?sourceMap'],
-			include: path.join(__dirname, 'app')
-		},  {
-			test: /\.(jpe?g|png|eot|woff|ttf|gif|svg)(\?.*)?$/i,
-			loader: 'file-loader',
-			include: path.join(__dirname, 'app')
 		}]
 	}
 };